+++ src/AdapterFactory/ReadOnlyAdapterFactory.php
@@ -2,22 +2,22 @@
 
 declare(strict_types=1);
 
-namespace Webf\Flysystem\Dsn;
+namespace Webf\Flysystem\Dsn\AdapterFactory;
 
-use League\Flysystem\PathPrefixing\PathPrefixedAdapter;
+use League\Flysystem\ReadOnly\ReadOnlyFilesystemAdapter;
 use Nyholm\Dsn\DsnParser;
 use Nyholm\Dsn\Exception\InvalidDsnException as NyholmInvalidDsnException;
 use Webf\Flysystem\Dsn\Exception\DsnException;
 use Webf\Flysystem\Dsn\Exception\DsnParameterException;
 use Webf\Flysystem\Dsn\Exception\UnsupportedDsnException;
 
-readonly class PrefixAdapterFactory implements FlysystemAdapterFactoryInterface
+readonly class ReadOnlyAdapterFactory implements FlysystemAdapterFactoryInterface
 {
     public function __construct(private FlysystemAdapterFactoryInterface $adapterFactory)
     {
     }
 
-    public function createAdapter(string $dsn): PathPrefixedAdapter
+    public function createAdapter(string $dsn): ReadOnlyFilesystemAdapter
     {
         $dsnString = $dsn;
         try {
@@ -26,21 +26,17 @@ readonly class PrefixAdapterFactory implements FlysystemAdapterFactoryInterface
             throw new DsnException($e->getMessage(), previous: $e);
         }
 
-        if ('prefix' !== $dsn->getName()) {
+        if ('readonly' !== $dsn->getName()) {
             throw UnsupportedDsnException::create($this, $dsnString);
         }
 
         if (count($arguments = $dsn->getArguments()) > 1) {
-            throw DsnParameterException::toManyArguments(1, count($arguments), 'prefix', $dsnString);
-        }
-
-        if (!is_string($path = $dsn->getParameter('path'))) {
-            throw DsnParameterException::missingParameter('path', $dsnString);
+            throw DsnParameterException::toManyArguments(1, count($arguments), 'readonly', $dsnString);
         }
 
         $innerAdapter = $this->adapterFactory->createAdapter($arguments[0]->__toString());
 
-        return new PathPrefixedAdapter($innerAdapter, $path);
+        return new ReadOnlyFilesystemAdapter($innerAdapter);
     }
 
     public function supports(string $dsn): bool
@@ -51,6 +47,6 @@ readonly class PrefixAdapterFactory implements FlysystemAdapterFactoryInterface
             throw new DsnException($e->getMessage(), previous: $e);
         }
 
-        return 'prefix' === $name;
+        return 'readonly' === $name;
     }
 }